+++ Scripts/bootstrap.sh
@@ -5,6 +5,7 @@
 #   - link git hooks
 #   - install required ruby gems
 #   - install sonar dependencies if `INSTALL_SONAR` environment variable is provided
+#   - install allure dependencies if `INSTALL_ALLURE` environment variable is provided
 # You should have homebrew installed.
 # If you get `zsh: permission denied: ./bootstrap.sh` error, please run `chmod +x bootstrap.sh` first
 
@@ -15,8 +16,8 @@ function puts {
 
 # Check if Homebrew is installed
 if [[ $(command -v brew) == "" ]]; then
-    echo "Homebrew not installed. Please install."
-    exit 1
+  echo "Homebrew not installed. Please install."
+  exit 1
 fi
 
 set -Eeuo pipefail
@@ -24,14 +25,13 @@ set -Eeuo pipefail
 trap "echo ; echo ❌ The Bootstrap script failed to finish without error. See the log above to debug. ; echo" ERR
 
 puts "Install Mint if needed"
-# List installed Mint versions, if fails, install Mint
-brew list mint || brew install mint
-
-# Set bash to Strict Mode (http://redsymbol.net/articles/unofficial-bash-strict-mode/)
-set -euo pipefail
+brew install mint
 
 puts "Bootstrap Mint dependencies"
 mint bootstrap
+
+# Set bash to Strict Mode (http://redsymbol.net/articles/unofficial-bash-strict-mode/)
+set -euo pipefail
 
 puts "Create git/hooks folder if needed"
 mkdir -p .git/hooks
@@ -47,12 +47,24 @@ chmod +x ./hooks/git-format-staged
 puts "Install bundle dependencies"
 bundle install
 
-if [[ ${INSTALL_SONAR-default} == true ]]; then
-    puts "Install sonar dependencies"
-    pip install lizard
-    brew install sonar-scanner
-fi
-
 # Copy internal Xcode scheme to the right folder for
 puts "Adding DemoApp-StreamDevelopers.xcscheme to the Xcode project"
 cp Scripts/DemoApp-StreamDevelopers.xcscheme StreamChat.xcodeproj/xcshareddata/xcschemes/DemoApp-StreamDevelopers.xcscheme
+
+if [[ ${INSTALL_SONAR-default} == true ]]; then
+  puts "Install sonar dependencies"
+  pip install lizard
+  brew install sonar-scanner
+fi
+
+if [[ ${INSTALL_ALLURE-default} == true ]]; then
+  puts "Install allurectl"
+  DOWNLOAD_URL="https://github.com/allure-framework/allurectl/releases/download/1.22.1/allurectl_darwin_amd64"
+  curl -sL "${DOWNLOAD_URL}" -o ./fastlane/allurectl
+  chmod +x ./fastlane/allurectl
+
+  puts "Install xcresults"
+  DOWNLOAD_URL="https://github.com/eroshenkoam/xcresults/releases/download/1.10.1/xcresults"
+  curl -sL "${DOWNLOAD_URL}" -o ./fastlane/xcresults
+  chmod +x ./fastlane/xcresults
+fi