+++ Sources/StreamChat/Database/DTOs/ChannelReadDTO.swift
@@ -7,7 +7,7 @@ import Foundation
 
 @objc(ChannelReadDTO)
 class ChannelReadDTO: NSManagedObject {
-    @NSManaged var lastReadAt: Date?
+    @NSManaged var lastReadAt: Date
     @NSManaged var unreadMessageCount: Int32
     
     // MARK: - Relationships
@@ -64,22 +64,6 @@ class ChannelReadDTO: NSManagedObject {
 
 extension NSManagedObjectContext {
     func saveChannelRead(
-        cid: ChannelId,
-        userId: UserId,
-        lastReadAt: Date,
-        unreadMessageCount: Int
-    ) -> ChannelReadDTO {
-        let dto = ChannelReadDTO.loadOrCreate(cid: cid, userId: userId, context: self)
-        
-        dto.user = UserDTO.loadOrCreate(id: userId, context: self)
-        
-        dto.lastReadAt = lastReadAt
-        dto.unreadMessageCount = Int32(unreadMessageCount)
-        
-        return dto
-    }
-    
-    func saveChannelRead(
         payload: ChannelReadPayload,
         for cid: ChannelId
     ) throws -> ChannelReadDTO {
@@ -95,13 +79,33 @@ extension NSManagedObjectContext {
     
     func markChannelAsRead(cid: ChannelId, userId: UserId, at: Date) {
         if let read = loadChannelRead(cid: cid, userId: userId) {
+            let previousLastReadAt = read.lastReadAt
+            
             // We have a read object saved, we can update it
             read.lastReadAt = at
             read.unreadMessageCount = 0
-        } else if let channel = channel(cid: cid), channel.members.contains(where: { $0.user.id == userId }) {
+            
+            // Mark messages authored by the current user sent within `previousLastReadAt...at` window
+            // as seen by the channel member with `userId`.
+            markMessagesFromCurrentUserAsRead(
+                for: read,
+                previousReadAt: previousLastReadAt
+            )
+        } else if let channel = channel(cid: cid), let member = channel.members.first(where: { $0.user.id == userId }) {
             // We don't have a read object, but the user is a member.
             // We can safely create a read object for the user
-            _ = saveChannelRead(cid: cid, userId: userId, lastReadAt: at, unreadMessageCount: 0)
+            let read = ChannelReadDTO.loadOrCreate(cid: cid, userId: userId, context: self)
+            read.channel = channel
+            read.user = member.user
+            read.lastReadAt = at
+            read.unreadMessageCount = 0
+            
+            // Mark all locally existed messages authored by the current user
+            // as seen by the channel member with `userId`.
+            markMessagesFromCurrentUserAsRead(
+                for: read,
+                previousReadAt: .distantPast
+            )
         } else {
             // If we don't have a read object saved for the user,
             // and the user is not a member,
@@ -112,6 +116,12 @@ extension NSManagedObjectContext {
             )
         }
     }
+    
+    func markChannelAsUnread(cid: ChannelId, by userId: UserId) {
+        guard let read = loadChannelRead(cid: cid, userId: userId) else { return }
+                
+        delete(read)
+    }
 
     func loadChannelRead(cid: ChannelId, userId: String) -> ChannelReadDTO? {
         ChannelReadDTO.load(cid: cid, userId: userId, context: self)
@@ -120,13 +130,34 @@ extension NSManagedObjectContext {
     func loadChannelReads(for userId: UserId) -> [ChannelReadDTO] {
         ChannelReadDTO.load(userId: userId, context: self)
     }
+    
+    private func markMessagesFromCurrentUserAsRead(
+        for read: ChannelReadDTO,
+        previousReadAt: Date
+    ) {
+        guard read.user.currentUser == nil else {
+            // Current user is not accounted in his own message reads.
+            return
+        }
+        
+        let messages = MessageDTO.loadCurrentUserMessages(
+            in: read.channel.cid,
+            createdAtFrom: previousReadAt,
+            createdAtThrough: read.lastReadAt,
+            context: self
+        )
+        
+        for message in messages {
+            message.reads.insert(read)
+        }
+    }
 }
 
 extension ChatChannelRead {
     fileprivate static func create(fromDTO dto: ChannelReadDTO) throws -> ChatChannelRead {
         guard dto.isValid else { throw InvalidModel(dto) }
         return try .init(
-            lastReadAt: dto.lastReadAt ?? Date.distantPast,
+            lastReadAt: dto.lastReadAt,
             unreadMessagesCount: Int(dto.unreadMessageCount),
             user: dto.user.asModel()
         )