+++ Sources/StreamChatUI/ChatChannelList/ChatChannelListItemView.swift
@@ -50,7 +50,7 @@ open class ChatChannelListItemView: _View, ThemeProvider, SwiftUIRepresentable {
 
     /// By default contains `subtitle` and `timestampLabel`.
     /// This container is embed inside `mainContainer ` and is the one below `topContainer`
-    open private(set) lazy var bottomContainer: ContainerStackView = ContainerStackView()
+    open private(set) lazy var bottomContainer: ContainerStackView = ContainerStackView(alignment: .center, spacing: 4)
         .withoutAutoresizingMaskConstraints
         .withAccessibilityIdentifier(identifier: "bottomContainer")
     
@@ -102,8 +102,18 @@ open class ChatChannelListItemView: _View, ThemeProvider, SwiftUIRepresentable {
         guard let content = content else { return nil }
         if let typingUsersInfo = typingUserString {
             return typingUsersInfo
-        } else if let latestMessage = content.channel.latestMessages.first {
-            return "\(latestMessage.author.name ?? latestMessage.author.id): \(latestMessage.textContent ?? latestMessage.text)"
+        } else if let previewMessage = content.channel.previewMessage {
+            guard previewMessage.type != .system else {
+                return previewMessage.text
+            }
+            
+            let authorName = previewMessage.isSentByCurrentUser
+                ? L10n.you
+                : previewMessage.author.name ?? previewMessage.author.id
+            
+            let text = previewMessage.textContent ?? previewMessage.text
+            
+            return "\(authorName): \(text)"
         } else {
             return L10n.Channel.Item.emptyMessages
         }
@@ -111,21 +121,38 @@ open class ChatChannelListItemView: _View, ThemeProvider, SwiftUIRepresentable {
 
     /// Text of `timestampLabel` which contains the time of the last sent message.
     open var timestampText: String? {
-        if let lastMessageAt = content?.channel.lastMessageAt {
-            return timestampFormatter.format(lastMessageAt)
+        if let timestamp = content?.channel.previewMessage?.createdAt {
+            return timestampFormatter.format(timestamp)
         } else {
             return nil
         }
     }
+    
+    /// The delivery status to be shown for the channel's preview message.
+    open var previewMessageDeliveryStatus: MessageDeliveryStatus? {
+        guard
+            let content = content,
+            let deliveryStatus = content.channel.previewMessage?.deliveryStatus
+        else { return nil }
+        
+        switch deliveryStatus {
+        case .pending, .failed:
+            return deliveryStatus
+        case .sent, .read:
+            guard content.channel.config.readEventsEnabled else { return nil }
+            
+            return deliveryStatus
+        default:
+            return nil
+        }
+    }
 
-    /*
-         TODO: ReadStatusView, Missing LLC API
-     /// The view showing indicator for read status of the last message in channel.
-     open private(set) lazy var readStatusView: ChatChannelReadStatusCheckmarkView = uiConfigSubviews
-         .readStatusView.init()
-         .withoutAutoresizingMaskConstraints
-      */
-
+    /// The indicator the delivery status of the channel preview message.
+    open private(set) lazy var previewMessageDeliveryStatusView = components
+        .messageDeliveryStatusCheckmarkView.init()
+        .withoutAutoresizingMaskConstraints
+        .withAccessibilityIdentifier(identifier: "previewMessageDeliveryStatusView")
+    
     override open func setUpAppearance() {
         super.setUpAppearance()
         backgroundColor = appearance.colorPalette.background
@@ -176,6 +203,8 @@ open class ChatChannelListItemView: _View, ThemeProvider, SwiftUIRepresentable {
         mainContainer.alignment = .center
         mainContainer.isLayoutMarginsRelativeArrangement = true
         
+        timestampLabel.setContentCompressionResistancePriority(.required, for: .horizontal)
+        
         embed(mainContainer)
     }
     
@@ -188,6 +217,19 @@ open class ChatChannelListItemView: _View, ThemeProvider, SwiftUIRepresentable {
 
         unreadCountView.content = content?.channel.unreadCount ?? .noUnread
         unreadCountView.invalidateIntrinsicContentSize()
+        
+        let checkmarkContent = previewMessageDeliveryStatus.map {
+            ChatMessageDeliveryStatusCheckmarkView.Content(deliveryStatus: $0)
+        }
+        previewMessageDeliveryStatusView.content = checkmarkContent
+        previewMessageDeliveryStatusView.isHidden = checkmarkContent == nil
+        
+        if let status = checkmarkContent?.deliveryStatus {
+            bottomContainer.insertArrangedSubview(
+                previewMessageDeliveryStatusView,
+                at: status == .pending || status == .failed ? 0 : 1
+            )
+        }
     }
 }